smsfilter/backend/predictor/pyproject.toml
[project]
name = "smsfilter-predictor"
version = "0.0.2.dev"
description = "Module for the SMS Filter predictor."
dependencies = [
    "torch>=2.9.1",
    "transformers>=4.57.6",
    "gguf>=0.17.1",
    "accelerate>=1.12.0",
    "llama-cpp-python>=0.3.16",
    "smsfilter-settings",
]

[tool.uv]
dev-dependencies = ["ruff>=0.12.1"]
package = true

[tool.uv.sources]
smsfilter-settings = { path = "../settings", editable = true }
torch = { index = "pytorch-cpu" }

[dependency-groups]
smsfilter_name_dependency_group = ["smsfilter-settings"]

[[tool.uv.index]]
name = "pytorch-cpu"
url = "https://download.pytorch.org/whl/cpu"
explicit = true

# Added by habushu-maven-plugin at 2026-01-18T17:29:11.190478461 to use https://pypi.org/simple/ as source repository for installing dependencies
[[tool.uv.index]]
name = "pypi"
url = "https://pypi.org/simple/"
publish-url = "https://pypi.org/simple/"
